requirements.txt
haystack-ai
chroma-haystack
pydantic==1.9.0